Lab 2. Adding H2 database

+++ sws/2023/src/main/java/io/pw/db/Product.java
@@ -3,5 +3,5 @@ package io.pw.db;
 /**
  * Created by pwykowski
  */
-public record Product(int id, String name, String desc, String serialNumber, int quantity) {
+public record Product(long id, String name, String desc, String serialNumber, int quantity) {
 }
